Add some files and reorder the folders
