+++ src/ProductsApp.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import {NavigationContainer} from '@react-navigation/native';
-import {StackNavigator} from './navigation/StackNavigator';
+import {StackNavigator} from './presentation/navigation/StackNavigator';
 
 export const ProductsApp = () => {
   return (
