+++ src/presentation/navigation/StackNavigator.tsx
@@ -1,0 +1,32 @@
+import React from 'react';
+import {createStackNavigator} from '@react-navigation/stack';
+import {LoadingScreen} from '../screens/loading/LoadingScreen';
+import {LoginScreen} from '../screens/auth/LoginScreen';
+import {RegisterScreen} from '../screens/auth/RegisterScreen';
+import {HomeScreen} from '../screens/home/HomeScreen';
+import {ProductScreen} from '../screens/product/ProductScreen';
+
+export type RootParams = {
+  LoadingScreen: undefined;
+  LoginScreen: undefined;
+  RegisterScreen: undefined;
+  HomeScreen: undefined;
+  ProductScreen: {productId: string};
+};
+
+const Stack = createStackNavigator<RootParams>();
+
+export const StackNavigator = () => {
+  return (
+    <Stack.Navigator
+      screenOptions={{
+        headerShown: false,
+      }}>
+      <Stack.Screen name="LoadingScreen" component={LoadingScreen} />
+      <Stack.Screen name="LoginScreen" component={LoginScreen} />
+      <Stack.Screen name="RegisterScreen" component={RegisterScreen} />
+      <Stack.Screen name="HomeScreen" component={HomeScreen} />
+      <Stack.Screen name="ProductScreen" component={ProductScreen} />
+    </Stack.Navigator>
+  );
+};
