+++ src/presentation/screens/auth/RegisterScreen.tsx
@@ -1,0 +1,10 @@
+import {View, Text} from 'react-native';
+import React from 'react';
+
+export const RegisterScreen = () => {
+  return (
+    <View>
+      <Text>RegisterScreen</Text>
+    </View>
+  );
+};
