+++ src/presentation/screens/loading/LoadingScreen.tsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import {Text, View} from 'react-native';
+
+export const LoadingScreen = () => {
+  return (
+    <View>
+      <Text>LoadingScreen</Text>
+    </View>
+  );
+};
