+++ src/presentation/screens/product/ProductScreen.tsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import {Text, View} from 'react-native';
+
+export const ProductScreen = () => {
+  return (
+    <View>
+      <Text>LoadingScreen</Text>
+    </View>
+  );
+};
